+++ src/REopt.jl
@@ -47,7 +47,8 @@ export
     backup_reliability,
     get_chp_defaults_prime_mover_size_class,
     get_steam_turbine_defaults_size_class,
-    simulated_load
+    simulated_load,
+    get_absorption_chiller_defaults
 
 import HTTP
 import JSON
@@ -82,6 +83,9 @@ const KWH_PER_MMBTU = 293.07107  # [kWh/mmbtu]
 const KWH_THERMAL_PER_TONHOUR = 3.51685
 const TONNE_PER_LB = 1/2204.62  # [tonne/lb]
 const FUEL_TYPES = ["natural_gas", "landfill_bio_gas", "propane", "diesel_oil"]
+const BIG_NUMBER = 1.0e10  #used for max size.  TODO use this number elsewhere.
+const PRIME_MOVERS = ["recip_engine", "micro_turbine", "combustion_turbine", "fuel_cell"]  #TODO replace `prime_movers` references in CHP code
+const HOT_WATER_OR_STEAM = ["steam", "hot_water"]  #TODO replace references to this list in chp, boiler
 const FUEL_DEFAULTS = Dict(
     "fuel_renewable_energy_fraction" => Dict(
         "natural_gas"=>0.0,
@@ -133,9 +137,9 @@ include("core/electric_load.jl")
 include("core/existing_boiler.jl")
 include("core/boiler.jl")
 include("core/existing_chiller.jl")
-include("core/absorption_chiller.jl")
 include("core/flexible_hvac.jl")
 include("core/heating_cooling_loads.jl")
+include("core/absorption_chiller.jl")
 include("core/electric_utility.jl")
 include("core/production_factor.jl")
 include("core/urdb.jl")